[processor] check whether spacing fits into MGRS tile boundaries
